+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # **Upcoming release**
 
 - #850 Update and pin black version in pre-commit and Github Actions
-- #850 Bump supported python version to up to Python 3.14
+- #851 Bump supported python version to up to Python 3.14
+- #852 Implement patchedast for TypeAlias
 
 # Release 1.14.0